% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-mechanisms.R
\name{find.good.L0}
\alias{find.good.L0}
\title{Finds an initial room-assignment vector}
\usage{
find.good.L0(N, M, A, n.draws = 10)
}
\arguments{
\item{N}{An integer. The number of units in the population.}

\item{M}{An integer. The number of units per room. `M` must be a divisor of
`N`.}

\item{A}{A numeric vector. The vector of attributes (must be of length `N`).}

\item{n.draws}{An integer. The number of potential room assignment vector to
draw, and from which to choose the best one.}
}
\value{
A numeric vector, representing the `best` room assignment vector among
those randomly generated.
}
\description{
Here, a `good' room-assignment vector is one whose corresponding exposure
vector is approximately balanced. The function generates `n.draws` potential
room assignment vectors and selects the one that is most balanced.
}
