% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frt.R
\name{FRT.sharp}
\alias{FRT.sharp}
\title{Runs a standard Fisher Randomization Test (for the sharp null hypothesis).}
\usage{
FRT.sharp(Zobs, Yobs, T, A, n.rand = 1000)
}
\arguments{
\item{Zobs}{A list of numeric vectors representing the group-assignment
structure.}

\item{Yobs}{A numeric vector of observed outcomes.}

\item{T}{A function representing the test statistic. It is called as
`T(W, Yobs)` in the code, where `W` is a vector of exposures, and `Yobs`
the vector of observed outcomes. IMPORTANT NOTE `T.dim` and `T.studentized`
cannot be used here, but `T.F` can.}

\item{A}{A numeric vector of attributes.}

\item{n.rand}{An integer. The number of draws to use to build the
null distribution of the test statistic (defaults to 1000).}
}
\description{
Runs a standard Fisher Randomization Test (for the sharp null hypothesis).
}
