% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{\%p\%}
\alias{\%p\%}
\title{Group action operator}
\usage{
s \%p\% v
}
\arguments{
\item{s}{A vector representing a permutation.}

\item{v}{A vector of the same length as \code{s}.}
}
\value{
A vector of the same length as \code{s} and \code{v}.
}
\description{
Applies a permutation \code{s} to a vector \code{v}.
}
\examples{
X <- seq(8)
A <- c(0, 0, 0, 0, 1, 1, 1, 1)
rsp(A) \%p\% X
rp(8) \%p\% X
}
