% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-mechanisms.R
\name{observe.outcomes}
\alias{observe.outcomes}
\title{Reveals the observed outcomes.}
\usage{
observe.outcomes(Z, A, science)
}
\arguments{
\item{Z}{A list of numeric vectors. A group-assignment structure.}

\item{A}{A numeric vector. A vector of attributes.}

\item{science}{A matrix. A full schedule of potential outcomes. Its number
of rows should equal the length of `A` and `Z`. Each column corresponds to
a different exposure. For now, exposures should be contiguous integers
starting from 0.}
}
\value{
A vector of observed outcomes, of the same length as `A` and `Z`.
}
\description{
Reveals the observed outcomes.
}
