% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{pprint}
\alias{pprint}
\title{Pretty print permutations}
\usage{
pprint(s)
}
\arguments{
\item{s}{A vector representing a permutation.}
}
\value{
A dataframe with 3 columns: \code{i}, \code{p.i}, \code{pm.i}.
}
\description{
\code{pprint} shows both the permutationa and its inverse. The column
labelled \code{p.i} shows the image of the column labelled \code{i} while
the image labelled \code{pm.i} shows the inverse image of \code{i}.
}
