% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-mechanisms.R
\name{rL}
\alias{rL}
\title{Generates one (or more) random vectors of room assignments}
\usage{
rL(N, M, n.draws = 10)
}
\arguments{
\item{N}{An integer. The total number of units in the population.}

\item{M}{An integer. The number of unit per room.}

\item{n.draws}{An integer. The number of room assignments vectors to
return (defaults to 10).}
}
\value{
If `n.draws = 1` then returns a vector of room assignments.
If `n.draws > 1', returns a matrix where each row represents a room
assignment vector.
}
\description{
Generates a room assignment vector (or a room assignment matrix)
depending on `n.draws`. For a vector, the entry at index `i`
represents the room to which unit `i` is to be assigned.
}
