% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{rp}
\alias{rp}
\title{Generates a random permutation}
\usage{
rp(x)
}
\arguments{
\item{x}{An integer or a vector.}
}
\value{
A vector representing a permutation.
}
\description{
If \code{x} is an integer, then returns an element of the symmetric
group on \code{x} elements. If \code{x} is a vector, then generates a 
permutation of the elements of \code{x}.
}
\examples{
rp(5)
rp(c(1, 4, 6, 10, 2))
}
