% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{rsp}
\alias{rsp}
\title{Generates a random stratified permutation}
\usage{
rsp(A)
}
\arguments{
\item{A}{A vector.}
}
\value{
A vector representing a permutation
}
\description{
\code{rsp} returns an element of the stabilizer of \code{A} in the
symmetric group on \code{|A|} elements.
}
\examples{
A <- c(0, 0, 0, 0, 1, 1, 1, 1)
rsp(A)
}
