% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-mechanisms.R
\name{rsW}
\alias{rsW}
\title{Generates a random exposure vector from a template W0}
\usage{
rsW(W0, A = NULL, n.draws = 10)
}
\arguments{
\item{W0}{An exposure vector.}

\item{A}{A vector of attributes. Should be of the same length as `W0`.}

\item{n.draws}{An integer. The number of permutations of `W0` that should be
drawn (defaults to 10).}
}
\value{
A matrix whose rows are random permutations of `W0`.
}
\description{
`rsW` draws a random permutation of `W0`. If `A` is null, then the
permutation will be drawn uniformly at random from the symmetric group
`S`. If A is not null, then the permutation will be drawn from `S_A`, the
stabilizer of `A` in the symmetric group.
}
