% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-mechanisms.R
\name{T.dim}
\alias{T.dim}
\title{Computes the difference-in-means test statistic}
\usage{
T.dim(W, Y, k, l, A = NULL)
}
\arguments{
\item{W}{A vector representing the assignment of units to treatments
(or exposures).}

\item{Y}{A vector representing the vector of observed outcomes. Must be
of the same length as \code{W}.}

\item{k}{An integer. One of the treatments to be contrasted.}

\item{l}{An integer. The other treatment to be contrasted.}

\item{A}{This parameter is ignored.}
}
\value{
A scalar.
}
\description{
\code{T.dim} computes the average difference in outcomes between the
units assigned to treatment \code{k} and those assigned to treatment
\code{l}.
}
\details{
The value of \code{A} is ignored. The parameter is here to allow a unified
treatment of test statistics by other functions.
}
