% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-mechanisms.R
\name{T.studentized}
\alias{T.studentized}
\title{Compute the stratified studentized test statistic.}
\usage{
T.studentized(W, Y, k, l, A)
}
\arguments{
\item{W}{A vector representing the assignment of units to treatments
(or exposures).}

\item{Y}{A vector representing the vector of observed outcomes. Must be
of the same length as \code{W}.}

\item{k}{An integer. One of the treatments to be contrasted.}

\item{l}{An integer. The other treatment to be contrasted.}

\item{A}{A vector representing the values to stratify on. Must be of
the same length as \code{A}.}
}
\value{
A scalar.
}
\description{
\code{T.studentizedd} computes the studentized difference in outcomes
between untis assigned to treatment \code{k} and those assigned to
treatment \code{l}, stratifying by values in \code{A}.
}
\details{
The test statistic is described in Basse et al. (2020+).
}
