% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-mechanisms.R
\name{W.fn}
\alias{W.fn}
\title{Computes the exposure vector associated with a group assignment.}
\usage{
W.fn(Z, A)
}
\arguments{
\item{Z}{A group assignment.}

\item{A}{An attribute vector.}
}
\value{
An exposure vector
}
\description{
Computes the exposure vector associated with a group assignment.
}
