% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-mechanisms.R
\name{W.star.fn}
\alias{W.star.fn}
\title{Computes the exposures associated with room assignments.}
\usage{
W.star.fn(L, A)
}
\arguments{
\item{L}{A room assignment vector or matrix.}

\item{A}{An attribute vector.}
}
\value{
If `L` is a vector, returns an exposure vector. If
`L` is a matrix, returns a matrix whose rows are exposure
vectors.
}
\description{
Computes the exposures associated with room assignments.
}
