% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-mechanisms.R
\name{Z.fn}
\alias{Z.fn}
\title{Converts a room assignment vector into a group assignment}
\usage{
Z.fn(L)
}
\arguments{
\item{L}{A room assignment vector.}
}
\value{
A list of numeric vector representing the corresponding
room assignment.
}
\description{
Converts a room assignment vector into a group assignment
}
